ajustes en componente piny of sale

+++ database/migrations/2025_12_20_141236_create_sale_items_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('sale_id')->constrained()->onDelete('cascade');
             $table->foreignId('product_id')->constrained()->onDelete('restrict');
-            $table->integer('quantity');
+            $table->decimal('quantity', 12, 2);
             $table->decimal('unit_price', 12, 2);
             $table->decimal('subtotal_local', 12, 2);
             $table->decimal('subtotal_usd', 12, 2);
